# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Venom K Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
...
  /courses/{slug}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course with chapters and lessons
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Admin get all users
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin update user (role, access)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Platform statistics
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 6
        languagePreference:
          type: string
          enum: [ar, en]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, username, email, role, xp, level, streak, languagePreference, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        languagePreference:
          type: string
          enum: [ar, en]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
        languagePreference:
          type: string
          enum: [ar, en]
        avatarUrl:
          type: string
    Course:
      type: object
      required: [id, slug, titleAr, titleEn, language, category, price, icon, color, totalLessons, totalChapters]
      properties:
        id:
          type: integer
        slug:
          type: string
        titleAr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: ["string", "null"]
        descriptionEn:
          type: ["string", "null"]
        language:
          type: string
        category:
          type: string
          enum: [programming, human_language]
        price:
          type: number
        icon:
          type: string
        color:
          type: string
        totalLessons:
          type: integer
        totalChapters:
          type: integer
        isFreeTrialAvailable:
          type: boolean
...
    Chapter:
      type: object
      required: [id, titleAr, titleEn, order, lessons]
      properties:
        id:
          type: integer
        titleAr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: ["string", "null"]
        descriptionEn:
          type: ["string", "null"]
        order:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/LessonPreview"
    LessonPreview:
      type: object
      required: [id, titleAr, titleEn, order, isFree, xpReward]
      properties:
        id:
          type: integer
        titleAr:
          type: string
        titleEn:
          type: string
        order:
          type: integer
        isFree:
          type: boolean
        xpReward:
          type: integer
        isCompleted:
          type: boolean
        hasVideo:
          type: boolean
    CourseDetail:
      type: object
      required: [id, slug, titleAr, titleEn, language, category, price, icon, color, chapters, totalLessons, totalChapters]
      properties:
        id:
          type: integer
        slug:
          type: string
        titleAr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: ["string", "null"]
        descriptionEn:
          type: ["string", "null"]
        language:
          type: string
        category:
          type: string
        price:
          type: number
        icon:
          type: string
        color:
          type: string
        totalLessons:
          type: integer
        totalChapters:
          type: integer
        isFreeTrialAvailable:
          type: boolean
        hasAccess:
          type: boolean
        userXpEarned:
          type: integer
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
...
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin]
        xp:
          type: integer
        level:
          type: integer
    AdminStats:
      type: object
      required: [totalUsers, totalCourses, totalLessons, totalSubscriptions, pendingSubscriptions, activeSubscriptions, totalXpAwarded, recentSignups]
      properties:
        totalUsers:
          type: integer
        totalCourses:
          type: integer
        totalLessons:
          type: integer
        totalSubscriptions:
          type: integer
        pendingSubscriptions:
          type: integer
        activeSubscriptions:
          type: integer
        totalXpAwarded:
          type: integer
        recentSignups:
          type: integer