# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Venom K Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { UserLanguagePreference } from './userLanguagePreference';
import type { UserRole } from './userRole';

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  xp: number;
  level: number;
  streak: number;
  languagePreference: UserLanguagePreference;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}